Refactor core and relocate Breadcrumbs
