+++ src/app/core/bread-crumb/bread-crumb.component.ts
@@ -1,7 +1,7 @@
 import {Component, ViewEncapsulation} from '@angular/core';
 import {ActivatedRoute, NavigationEnd, Router} from '@angular/router';
-import {ErrorLogService, XmppService} from '../../core';
-import {BreadCrumb, BreadCrumbs} from '../../models';
+import {ErrorLogService, XmppService} from '../';
+import {BreadCrumb, BreadCrumbs} from './';
 import {Observable} from 'rxjs/Observable';
 import 'rxjs/add/operator/distinctUntilChanged';
 import 'rxjs/add/operator/map';
