+++ src/app/core/bread-crumb/index.ts
@@ -1,1 +1,2 @@
+export * from './bread-crumb';
 export * from './bread-crumb.component';
