+++ src/app/core/navigation/navigation.service.ts
@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {Router} from '@angular/router';
-import {Topic} from '../models';
+import {Topic} from '../../models';
 
 @Injectable()
 export class NavigationService {
