+++ src/app/core/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './custom-url-serializer';
+export * from './global-error-handler.service';
